refactor: change enum file location batch
