+++ src/domain/auth/constant/platform.const.ts
